@@ -1,27 +1,29 @@
 import { Setting } from './setting'
-import axios from './http'
+import { http } from './http'
 class BookmarkService {
     async get() {
         let setting = await Setting.build();
-        let resp = await axios.get(`/gists/${setting.gistID}`)
-        if (resp && resp.data) {
-            let filenames = Object.keys(resp.data.files);
+        let resp = await http.get(`gists/${setting.gistID}`).json() as any
+        if (resp?.files) {
+            let filenames = Object.keys(resp.files);
             if (filenames.indexOf(setting.gistFileName) !== -1) {
-                let gistFile = resp.data.files[setting.gistFileName]
+                let gistFile = resp.files[setting.gistFileName]
                 if (gistFile.truncated) {
-                    return axios.get(gistFile.raw_url, { responseType: 'blob' }).then(resp => resp.data.text())
+                    const txt = http.get(gistFile.raw_url).text();
+                    return txt;
                 } else {
                     return gistFile.content
                 }
             }
         }
+        return resp;
     }
     async getAllGist() {
-        return axios.get('/gists').then(resp => resp.data)
+        return http.get('gists').json();
     }
     async update(data: any) {
         let setting = await Setting.build();
-        return axios.patch(`/gists/${setting.gistID}`, data).then(resp => resp.data)
+        return http.patch(`gists/${setting.gistID}`, { json: data }).json();
     }
 }
 
